requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.13
django-behaviors==0.5.1
django-filter==21.1
django-lifecycle==1.0.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
docutils==0.18.1
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
parse==1.19.0
parse-type==0.6.0
Pillow==9.1.1
psycopg2-binary==2.9.3
Pygments==2.12.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.1
requests==2.28.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
urlman==2.0.1
whitenoise==3.3.1
zipp==3.8.0
